Exercise 4.21*: bloglist expansion, step9 completed

+++ controllers/blogs.js
@@ -51,8 +51,19 @@ blogsRouter.post("/", async (request, response, next) => {
 // Delete blog item
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      return response
+        .status(401)
+        .json({ error: "you are unauthorized to perfom that action" });
+    }
   } catch (exception) {
     next(exception);
   }
